Fix dtype on Keyfeature. (#129)

+++ tests/integration/butterfree/core/transform/test_aggregated_feature_set.py
@@ -232,7 +232,13 @@ class TestAggregatedFeatureSet:
                     ).with_pivot("pivot_col", ["S", "N"]),
                 )
             ],
-            keys=[KeyFeature(name="id", description="The user's Main ID or device ID")],
+            keys=[
+                KeyFeature(
+                    name="id",
+                    description="The user's Main ID or device ID",
+                    dtype=DataType.INTEGER,
+                )
+            ],
             timestamp=TimestampFeature(),
         )
